backend/orders_service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic[email]==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
alembic==1.13.1
